# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Developer Multi-Agent Workflow API
  version: 0.1.0
  description: Internal API for developer workflow planning, review, and test strategy generation.
servers:
  - url: http://localhost:8000
tags:
  - name: workflows
  - name: feedback
paths:
  /health:
    get:
      tags: [workflows]
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /v1/workflows/plan:
    post:
      tags: [workflows]
      summary: Create a development plan
      operationId: createWorkflowPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanRequest"
      responses:
        "200":
          description: Completed plan response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"

  /v1/workflows/review:
    post:
      tags: [workflows]
      summary: Review a proposed change or PR
      operationId: createWorkflowReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
      responses:
        "200":
          description: Completed review response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponse"

  /v1/workflows/test-plan:
    post:
      tags: [workflows]
      summary: Generate a test strategy
      operationId: createTestPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestPlanRequest"
      responses:
        "200":
          description: Completed test plan response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestPlanResponse"

  /v1/workflows/{run_id}:
    get:
      tags: [workflows]
      summary: Get workflow run by ID
      operationId: getWorkflowRun
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunDetailResponse"
        "404":
          description: Run not found

  /v1/workflows/{run_id}/trace:
    get:
      tags: [workflows]
      summary: Get workflow trace summary
      operationId: getWorkflowTrace
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceResponse"

  /v1/feedback:
    post:
      tags: [feedback]
      summary: Store workflow feedback
      operationId: createFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "200":
          description: Feedback stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
      required: [status]

    Evidence:
      type: object
      properties:
        source_type:
          type: string
          example: repo
        source_id:
          type: string
          nullable: true
        locator:
          type: string
          example: src/profile/service.ts
        snippet:
          type: string
          nullable: true
        timestamp:
          type: string
          nullable: true
        confidence:
          type: string
          enum: [low, medium, high]
      required: [source_type, locator, confidence]

    ArtifactRefs:
      type: object
      properties:
        issue_ids:
          type: array
          items:
            type: string
        pr_url:
          type: string
          nullable: true
        changed_files:
          type: array
          items:
            type: string

    RequestOptions:
      type: object
      properties:
        include_tests:
          type: boolean
          default: true
        language:
          type: string
          default: ko

    BaseWorkflowRequest:
      type: object
      properties:
        request_type:
          type: string
          enum: [feature, bugfix, refactor, review, test_plan]
        repo_id:
          type: string
        branch:
          type: string
        task_text:
          type: string
        artifacts:
          $ref: "#/components/schemas/ArtifactRefs"
        options:
          $ref: "#/components/schemas/RequestOptions"
      required: [request_type, repo_id, branch, task_text]

    PlanRequest:
      allOf:
        - $ref: "#/components/schemas/BaseWorkflowRequest"

    ReviewRequest:
      allOf:
        - $ref: "#/components/schemas/BaseWorkflowRequest"
        - type: object
          properties:
            diff_text:
              type: string
              nullable: true

    TestPlanRequest:
      type: object
      properties:
        repo_id:
          type: string
        branch:
          type: string
        implementation_plan:
          type: array
          items:
            type: string
        impacted_areas:
          type: array
          items:
            type: string
      required: [repo_id, branch, implementation_plan, impacted_areas]

    BaseWorkflowResponse:
      type: object
      properties:
        run_id:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        trace_id:
          type: string
        request_type:
          type: string
        confidence:
          type: string
          enum: [low, medium, high]
      required: [run_id, status, trace_id, confidence]

    PlanResponse:
      allOf:
        - $ref: "#/components/schemas/BaseWorkflowResponse"
        - type: object
          properties:
            summary:
              type: string
            impacted_areas:
              type: array
              items:
                type: string
            implementation_plan:
              type: array
              items:
                type: string
            tests:
              type: array
              items:
                type: string
            risks:
              type: array
              items:
                type: string
            open_questions:
              type: array
              items:
                type: string
            evidence:
              type: array
              items:
                $ref: "#/components/schemas/Evidence"
          required:
            - summary
            - impacted_areas
            - implementation_plan
            - tests
            - risks
            - open_questions
            - evidence

    ReviewFinding:
      type: object
      properties:
        category:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        message:
          type: string
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
      required: [category, severity, message]

    ReviewResponse:
      allOf:
        - $ref: "#/components/schemas/BaseWorkflowResponse"
        - type: object
          properties:
            summary:
              type: string
            review_findings:
              type: array
              items:
                $ref: "#/components/schemas/ReviewFinding"
            missing_tests:
              type: array
              items:
                type: string
            risks:
              type: array
              items:
                type: string
            readiness_verdict:
              type: string
              enum: [ready, needs_changes, blocked]
            evidence:
              type: array
              items:
                $ref: "#/components/schemas/Evidence"
          required:
            - summary
            - review_findings
            - missing_tests
            - risks
            - readiness_verdict
            - evidence

    TestPlanResponse:
      allOf:
        - $ref: "#/components/schemas/BaseWorkflowResponse"
        - type: object
          properties:
            unit_tests:
              type: array
              items:
                type: string
            integration_tests:
              type: array
              items:
                type: string
            regression_targets:
              type: array
              items:
                type: string
            edge_cases:
              type: array
              items:
                type: string
            execution_order:
              type: array
              items:
                type: string
          required:
            - unit_tests
            - integration_tests
            - regression_targets
            - edge_cases
            - execution_order

    RunDetailResponse:
      type: object
      properties:
        run_id:
          type: string
        status:
          type: string
        created_at:
          type: string
        completed_at:
          type: string
          nullable: true
        request:
          type: object
        result:
          type: object
        trace_id:
          type: string
      required: [run_id, status, created_at, request, trace_id]

    TraceStep:
      type: object
      properties:
        step_name:
          type: string
        status:
          type: string
        latency_ms:
          type: integer
        tool_calls:
          type: integer
        error_message:
          type: string
          nullable: true
      required: [step_name, status, latency_ms, tool_calls]

    TraceResponse:
      type: object
      properties:
        trace_id:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TraceStep"
        error_summary:
          type: string
          nullable: true
      required: [trace_id, steps]

    FeedbackRequest:
      type: object
      properties:
        run_id:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        useful:
          type: boolean
        comment:
          type: string
          nullable: true
      required: [run_id, rating, useful]

    FeedbackResponse:
      type: object
      properties:
        feedback_id:
          type: string
        stored:
          type: boolean
      required: [feedback_id, stored]